Add LINQPractise1 project with LINQ example
A new project named `LINQPractise1` has been added to the Visual Studio solution file `LINQPractise.sln`. This includes the project configuration for both Debug and Release modes.

A new project file `LINQPractise1.csproj` has been created for the `LINQPractise1` project. This project targets .NET 8.0 and has implicit usings and nullable reference types enabled.

A new `Program.cs` file has been added to the `LINQPractise1` project. This file contains a C# program that uses LINQ to find positive numbers from a list of integers that are within the range of 1 to 11. The program outputs these numbers to the console.